Refactoring Session updates part
